# Conflicts:
#	Mashd.Application/MashdInterpreter.cs
#	Mashd.Frontend/AST/AstBuilder.cs
#	Mashd.Frontend/AST/Definitions/DatasetDefinitionNode.cs
#	Mashd.Frontend/AST/Definitions/FormalParameterNode.cs
#	Mashd.Frontend/AST/Expressions/IdentifierNode.cs
#	Mashd.Frontend/SemanticAnalysis/TypeChecker.cs
#	Mashd.Test/ImportUnitTests.cs

+++ Mashd.Frontend/AST/AstBuilder.cs
@@ -446,8 +446,8 @@ public class AstBuilder : MashdBaseVisitor<AstNode>
         var (line, column, text) = ExtractNodeInfo(context);
 
         string intText = context.INTEGER().GetText();
-
-        int value = int.Parse(intText);
+        
+        long value = long.Parse(intText);
         return new LiteralNode(value, line, column, text, SymbolType.Integer);
     }
 
@@ -458,7 +458,7 @@ public class AstBuilder : MashdBaseVisitor<AstNode>
 
         string decText = context.DECIMAL().GetText();
 
-        decimal value = decimal.Parse(decText);
+        double value = double.Parse(decText);
         return new LiteralNode(value, line, column, text, SymbolType.Decimal);
     }
